@@ -53,11 +53,11 @@
     "postcss-loader": "2.1.5",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "react": "16.4.2",
+    "react": "^16.7.0-alpha.2",
     "react-dev-utils": "6.0.0-next.a671462c",
     "react-dnd": "^5.0.0",
     "react-dnd-touch-backend": "^0.5.1",
-    "react-dom": "16.4.2",
+    "react-dom": "^16.7.0-alpha.2",
     "react-ga": "^2.5.3",
     "react-router-dom": "4.3.1",
     "resolve": "1.6.0",
