@@ -1,24 +1,18 @@
-import React from 'react';
-class ComponentLoader extends React.Component {
-  state = {
+import React, { useEffect, useState } from "react";
+
+function ComponentLoader(props) {
+  const [state, setState] = useState({
     Component: null
-  };
-  async UNSAFE_componentWillMount() {
-    try {
-      await this.fetchComponent();
-    } catch (e) {
-      console.log(e);
-    }
-  }
-  async fetchComponent() {
-    const Component = await this.props.loader();
-    this.setState({
-      Component: Component && Component.default ? Component.default : Component
+  });
+  useEffect(() => {
+    props.loader().then(Component => {
+      setState({
+        Component:
+          Component && Component.default ? Component.default : Component
+      });
     });
-  }
-  render() {
-    const { Component } = this.state;
-    return Component ? <Component {...this.props} /> : null;
-  }
+  }, []);
+  const { Component } = state;
+  return Component ? <Component {...props} /> : null;
 }
 export default ComponentLoader;
