@@ -1,43 +1,48 @@
-import React,{ Component } from "react";
-
-
-import { AudioConsumer } from "./../AudioManager";
-import PieceDropTarget from "./PieceDropTarget";
-import Piece from "./Piece";
 import "../../styles/picture.scss";
 
-class Picture extends Component {
-  render() {
-    const { size, pieces, picture, dropPiece } = this.props;
-    const styles = {
-      width: size.width,
-      height: size.height,
-      backgroundImage: `url(${picture})`
-    };
-    return <div className="container picture-wrapper">
-        <div className="picture-bg">
-          <div className="picture" style={styles}>
-            {pieces.map((piece, key) => {
-              return <AudioConsumer key={key}>
-                  {({ methods }) => {
-                    const style = { width: piece.size, height: piece.size, top: piece.y, left: piece.x };
-                    const baseProps = { style, piece, dropPiece, playDrop:methods.playDrop };
-                    return piece.matched ? <Piece {...baseProps} /> : <PieceDropTarget
-                        {...{ style, ...baseProps }}
-                      >
-                        {extra => (
-                          <Piece
-                            {...{ ...baseProps, ...extra }}
-                          />
-                        )}
-                      </PieceDropTarget>;
-                  }}
-                </AudioConsumer>;
-            })}
-          </div>
+import React, { useContext } from "react";
+
+import { AudioContext } from "./../AudioManager";
+import Piece from "./Piece";
+import PieceDropTarget from "./PieceDropTarget";
+
+function Picture(props) {
+  const { methods } = useContext(AudioContext);
+  const { size, pieces, picture, dropPiece } = props;
+  const styles = {
+    width: size.width,
+    height: size.height,
+    backgroundImage: `url(${picture})`
+  };
+  return (
+    <div className="container picture-wrapper">
+      <div className="picture-bg">
+        <div className="picture" style={styles}>
+          {pieces.map((piece, key) => {
+            const style = {
+              width: piece.size,
+              height: piece.size,
+              top: piece.y,
+              left: piece.x
+            };
+            const baseProps = {
+              style,
+              piece,
+              dropPiece,
+              playDrop: methods.playDrop
+            };
+            return piece.matched ? (
+              <Piece key={key} {...baseProps} />
+            ) : (
+              <PieceDropTarget key={key} {...{ style, ...baseProps }}>
+                {extra => <Piece {...{ ...baseProps, ...extra }} />}
+              </PieceDropTarget>
+            );
+          })}
         </div>
-      </div>;
-  }
+      </div>
+    </div>
+  );
 }
 
 export default Picture;
